+++ crate/target/debug/examples/p-9ce19d432ef72989.d
@@ -1,0 +1,5 @@
+/Users/dev/dev/extensions/numbers-le/crate/target/debug/examples/p-9ce19d432ef72989.d: examples/p.rs
+
+/Users/dev/dev/extensions/numbers-le/crate/target/debug/examples/p-9ce19d432ef72989: examples/p.rs
+
+examples/p.rs: