+++ crate/target/debug/examples/p.d
@@ -1,0 +1,1 @@
+/Users/dev/dev/extensions/numbers-le/crate/target/debug/examples/p: /Users/dev/dev/extensions/numbers-le/crate/examples/p.rs